backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
Pillow>=10.0
python-dotenv>=1.0
dj-database-url>=2.0
psycopg2-binary>=2.9
gunicorn>=21.0
